+++ Sources/Utils/CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.36.0 (pre-release):
+* [Fix] Fix handling the negative values in all of the type formatters. E.g. [`GUIUtils.AngleType`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_AngleType.htm).
 * [Enhancement] Auto update modules that implement [`GUIUtils.IHasContextMenu`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_IHasContextMenu.htm).
 * [Enhancement] Support localization for the `UI_Toggle` attribute: [`GUIUtils.StdSpecTags](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_StdSpecTags.htm).
 * [Enhancement] On language version change, update the events and menus even when the part doesn't have a config: [`GUIUtils.LocalizationLoader](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_LocalizationLoader.htm).